unificar módulos para simplificar estrutura do pacote `core.data`
