+++ experimentos/experiments/core/data/__init__.py
@@ -9,17 +9,27 @@ This package provides a pipeline-based architecture for data processing:
 from .corporate_credit import corporate_credit_transformer
 from .datasets import Dataset
 from .lending_club import lending_club_transformer
-from .registry import get_transformer, get_transformer_registry, register_transformer
+from .repository import DataRepository
 from .taiwan_credit import taiwan_credit_transformer
-from .transformer import Transformer
+from .transformers import (
+    Transformer,
+    TransformerRegistry,
+    get_engine,
+    get_transformer,
+    get_transformer_registry,
+    register_transformer,
+)
 
 __all__ = [
     "Dataset",
+    "DataRepository",
     "Transformer",
+    "TransformerRegistry",
+    "get_transformer",
+    "get_transformer_registry",
+    "get_engine",
     "corporate_credit_transformer",
     "lending_club_transformer",
     "taiwan_credit_transformer",
     "register_transformer",
-    "get_transformer",
-    "get_transformer_registry",
 ]
