+++ experimentos/experiments/core/data/lending_club.py
@@ -7,8 +7,8 @@ import re
 import polars as pl
 from polars import datatypes
 
-from experiments.core.data.registry import register_transformer
-from experiments.core.data.transformer import get_engine
+from experiments.core.data.datasets import Dataset
+from experiments.core.data.transformers import get_engine, register_transformer
 
 
 @register_transformer("lending_club")
